#28469 Fixing more issues.

+++ Tests/Caravela.Framework.Tests.Integration/Tests/Templating/LocalVariables/CompileTimeIsPatternVariable.t.cs
@@ -1,3 +1,3 @@
 {
-    return this.Method(a);
+    return (int)this.Method(a);
 }
